requirements.txt
alembic==1.0.0
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
click==6.7
Flask==1.1.dev0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.6
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
flaskr==0.0.0
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.3
psycopg2==2.7.5
pycparser==2.18
PyJWT==1.6.4
python-dateutil==2.7.3
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.10
virtualenv==16.0.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
WTForms==2.2.1
